+++ tests/integration/services/test_service_factory.py
@@ -19,34 +19,6 @@ import pytest
 from craft_application import services
 
 
-def test_gets_dataclass_services(
-    check,
-    app_metadata,
-    fake_project,
-    project_path,
-    fake_package_service_class,
-    fake_lifecycle_service_class,
-    fake_project_service_class,
-    fake_provider_service_class,
-):
-    with pytest.warns(DeprecationWarning, match="Use ServiceFactory.register"):
-        factory = services.ServiceFactory(
-            app_metadata,
-            project=fake_project,
-            PackageClass=fake_package_service_class,
-            LifecycleClass=fake_lifecycle_service_class,
-            ProjectClass=fake_project_service_class,
-            ProviderClass=fake_provider_service_class,
-        )
-    factory.update_kwargs("project", project_dir=project_path)
-    factory.get("project").configure(platform=None, build_for=None)
-    factory.get("project").set(fake_project)  # type: ignore[reportAttributeAccessIssue]
-
-    check.is_instance(factory.package, services.PackageService)
-    check.is_instance(factory.lifecycle, services.LifecycleService)
-    check.is_instance(factory.provider, services.ProviderService)
-
-
 def test_gets_registered_services(
     check,
     app_metadata,
@@ -74,9 +46,9 @@ def test_gets_registered_services(
     check.is_instance(factory.get("provider"), services.ProviderService)
 
 
-def test_real_service_error(app_metadata, fake_project):
+def test_real_service_error(app_metadata):
     services.ServiceFactory.register("package", services.PackageService)
-    factory = services.ServiceFactory(app_metadata, project=fake_project)
+    factory = services.ServiceFactory(app_metadata)
 
     with pytest.raises(
         TypeError,